+++ packages/indexer-api/src/controllers/index.ts
@@ -1,1 +1,2 @@
-export * from "./indexer";
+export * from "./balances";
+export * from "./deposits";
